+++ source/DSCResources/DSC_DnsServerSetting/DSC_DnsServerSetting.psm1
@@ -6,7 +6,7 @@ Import-Module -Name $script:dnsServerDscCommonPath
 
 $script:localizedData = Get-LocalizedData -DefaultUICulture 'en-US'
 
-$properties = 'LocalNetPriority', 'AutoConfigFileZones', 'MaxCacheTTL', 'AddressAnswerLimit', 'UpdateOptions', 'DisableAutoReverseZones', 'StrictFileParsing', 'NoRecursion', 'DisjointNets', 'EnableDirectoryPartitions', 'XfrConnectTimeout', 'AllowUpdate', 'DsAvailable', 'BootMethod', 'LooseWildcarding', 'DsPollingInterval', 'BindSecondaries', 'LogLevel', 'AutoCacheUpdate', 'EnableDnsSec', 'EnableEDnsProbes', 'NameCheckFlag', 'EDnsCacheTimeout', 'SendPort', 'WriteAuthorityNS', 'IsSlave', 'RecursionTimeout', 'ListenAddresses', 'DsTombstoneInterval', 'RecursionRetry', 'RpcProtocol', 'SecureResponses', 'RoundRobin', 'ForwardDelegations', 'MaxNegativeCacheTTL'
+$properties = 'LocalNetPriority', 'AutoConfigFileZones', 'AddressAnswerLimit', 'UpdateOptions', 'DisableAutoReverseZones', 'StrictFileParsing', 'NoRecursion', 'DisjointNets', 'EnableDirectoryPartitions', 'XfrConnectTimeout', 'AllowUpdate', 'DsAvailable', 'BootMethod', 'LooseWildcarding', 'DsPollingInterval', 'BindSecondaries', 'LogLevel', 'AutoCacheUpdate', 'EnableDnsSec', 'NameCheckFlag', 'SendPort', 'WriteAuthorityNS', 'IsSlave', 'RecursionTimeout', 'ListenAddresses', 'DsTombstoneInterval', 'RecursionRetry', 'RpcProtocol', 'RoundRobin', 'ForwardDelegations'
 
 <#
     .SYNOPSIS
@@ -90,10 +90,6 @@ function Get-TargetResource
         Lifetime of tombstoned records in Directory Service integrated zones,
         expressed in seconds.
 
-    .PARAMETER EDnsCacheTimeout
-        Lifetime, in seconds, of the cached information describing the EDNS version
-        supported by other DNS Servers.
-
     .PARAMETER EnableDirectoryPartitions
         Specifies whether support for application directory partitions is enabled on
         the DNS Server.
@@ -101,13 +97,6 @@ function Get-TargetResource
     .PARAMETER EnableDnsSec
         Specifies whether the DNS Server includes DNSSEC-specific RRs, KEY, SIG, and
         NXT in a response.
-
-    .PARAMETER EnableEDnsProbes
-        Specifies the behavior of the DNS Server. When TRUE, the DNS Server always
-        responds with OPT resource records according to RFC 2671, unless the remote
-        server has indicated it does not support EDNS in a prior exchange. If FALSE,
-        the DNS Server responds to queries with OPTs only if OPTs are sent in the
-        original query.
 
     .PARAMETER ForwardDelegations
         Specifies whether queries to delegated sub-zones are forwarded.
@@ -130,14 +119,6 @@ function Get-TargetResource
     .PARAMETER LooseWildcarding
         Indicates whether the DNS Server performs loose wildcarding.
 
-    .PARAMETER MaxCacheTTL
-        Maximum time, in seconds, the record of a recursive name query may remain
-        in the DNS Server cache.
-
-    .PARAMETER MaxNegativeCacheTTL
-        Maximum time, in seconds, a name error result from a recursive query may
-        remain in the DNS Server cache.
-
     .PARAMETER NameCheckFlag
         Indicates the set of eligible characters to be used in DNS names.
 
@@ -156,10 +137,6 @@ function Get-TargetResource
 
     .PARAMETER RpcProtocol
         RPC protocol or protocols over which administrative RPC runs.
-
-    .PARAMETER SecureResponses
-        Indicates whether the DNS Server exclusively saves records of names in the
-        same subtree as the server that provided them.
 
     .PARAMETER SendPort
         Port on which the DNS Server sends UDP queries to other servers.
@@ -229,20 +206,12 @@ function Set-TargetResource
         $DsTombstoneInterval,
 
         [Parameter()]
-        [uint32]
-        $EDnsCacheTimeout,
-
-        [Parameter()]
         [bool]
         $EnableDirectoryPartitions,
 
         [Parameter()]
         [uint32]
         $EnableDnsSec,
-
-        [Parameter()]
-        [bool]
-        $EnableEDnsProbes,
 
         [Parameter()]
         [uint32]
@@ -270,14 +239,6 @@ function Set-TargetResource
 
         [Parameter()]
         [uint32]
-        $MaxCacheTTL,
-
-        [Parameter()]
-        [uint32]
-        $MaxNegativeCacheTTL,
-
-        [Parameter()]
-        [uint32]
         $NameCheckFlag,
 
         [Parameter()]
@@ -299,10 +260,6 @@ function Set-TargetResource
         [Parameter()]
         [int16]
         $RpcProtocol,
-
-        [Parameter()]
-        [bool]
-        $SecureResponses,
 
         [Parameter()]
         [uint32]
@@ -404,10 +361,6 @@ function Set-TargetResource
         Lifetime of tombstoned records in Directory Service integrated zones,
         expressed in seconds.
 
-    .PARAMETER EDnsCacheTimeout
-        Lifetime, in seconds, of the cached information describing the EDNS version
-        supported by other DNS Servers.
-
     .PARAMETER EnableDirectoryPartitions
         Specifies whether support for application directory partitions is enabled on
         the DNS Server.
@@ -415,13 +368,6 @@ function Set-TargetResource
     .PARAMETER EnableDnsSec
         Specifies whether the DNS Server includes DNSSEC-specific RRs, KEY, SIG, and
         NXT in a response.
-
-    .PARAMETER EnableEDnsProbes
-        Specifies the behavior of the DNS Server. When TRUE, the DNS Server always
-        responds with OPT resource records according to RFC 2671, unless the remote
-        server has indicated it does not support EDNS in a prior exchange. If FALSE,
-        the DNS Server responds to queries with OPTs only if OPTs are sent in the
-        original query.
 
     .PARAMETER ForwardDelegations
         Specifies whether queries to delegated sub-zones are forwarded.
@@ -444,14 +390,6 @@ function Set-TargetResource
     .PARAMETER LooseWildcarding
         Indicates whether the DNS Server performs loose wildcarding.
 
-    .PARAMETER MaxCacheTTL
-        Maximum time, in seconds, the record of a recursive name query may remain
-        in the DNS Server cache.
-
-    .PARAMETER MaxNegativeCacheTTL
-        Maximum time, in seconds, a name error result from a recursive query may
-        remain in the DNS Server cache.
-
     .PARAMETER NameCheckFlag
         Indicates the set of eligible characters to be used in DNS names.
 
@@ -470,10 +408,6 @@ function Set-TargetResource
 
     .PARAMETER RpcProtocol
         RPC protocol or protocols over which administrative RPC runs.
-
-    .PARAMETER SecureResponses
-        Indicates whether the DNS Server exclusively saves records of names in the
-        same subtree as the server that provided them.
 
     .PARAMETER SendPort
         Port on which the DNS Server sends UDP queries to other servers.
@@ -544,20 +478,12 @@ function Test-TargetResource
         $DsTombstoneInterval,
 
         [Parameter()]
-        [uint32]
-        $EDnsCacheTimeout,
-
-        [Parameter()]
         [bool]
         $EnableDirectoryPartitions,
 
         [Parameter()]
         [uint32]
         $EnableDnsSec,
-
-        [Parameter()]
-        [bool]
-        $EnableEDnsProbes,
 
         [Parameter()]
         [uint32]
@@ -585,14 +511,6 @@ function Test-TargetResource
 
         [Parameter()]
         [uint32]
-        $MaxCacheTTL,
-
-        [Parameter()]
-        [uint32]
-        $MaxNegativeCacheTTL,
-
-        [Parameter()]
-        [uint32]
         $NameCheckFlag,
 
         [Parameter()]
@@ -614,10 +532,6 @@ function Test-TargetResource
         [Parameter()]
         [int16]
         $RpcProtocol,
-
-        [Parameter()]
-        [bool]
-        $SecureResponses,
 
         [Parameter()]
         [uint32]